create lab1
