--- WindowsFormsCars/WindowsFormsCars/Plane.cs
@@ -1,101 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Drawing;
-
-namespace WindowsFormsPlane
-{
-    /// <summary>
-    /// Класс отрисовки самолета
-    /// </summary>
-    public class Plane : APlane
-    {
-        /// <summary>
-        /// Ширина отрисовки самолета
-        /// </summary>
-        private readonly int planeWidth = 150;
-
-        /// <summary>
-        /// Высота отрисовки самолета
-        /// </summary>
-        private readonly int planeHeight = 120;
-
-        /// <summary>
-        /// Конструктор
-        /// </summary>
-        /// <param name="maxSpeed">Максимальная скорость</param>
-        /// <param name="weight">Вес автомобиля</param>
-        /// <param name="mainColor">Основной цвет кузова</param>
-        /// <param name="dopColor">Дополнительный цвет</param>
-        /// <param name="frontSpoiler">Признак наличия переднего спойлера</param>
-        /// <param name="sideSpoiler">Признак наличия боковых спойлеров</param>
-        /// <param name="backSpoiler">Признак наличия заднего спойлера</param>
-        /// <param name="sportLine">Признак наличия гоночной полосы</param>
-        public Plane(int maxSpeed, float weight, Color mainColor)
-        {
-            MaxSpeed = maxSpeed;
-            Weight = weight;
-            MainColor = mainColor;
-        }
-
-        /// <summary>
-        /// Изменение направления пермещения
-        /// </summary>
-        /// <param name="direction">Направление</param>
-        public override void MoveTransport(Direction direction)
-        {
-            float step = MaxSpeed * 100 / Weight;
-            switch (direction)
-            {
-                // вправо
-                case Direction.Right:
-                    if (_startPosX + step < _pictureWidth - planeWidth)
-                    {
-                        _startPosX += step;
-                    }
-                    break;
-                //влево
-                case Direction.Left:
-                    if (_startPosX - step > 0)
-                    {
-                        _startPosX -= step;
-                    }
-                    break;
-                //вверх
-                case Direction.Up:
-                    if (_startPosY - step > 0)
-                    {
-                        _startPosY -= step;
-                    }
-                    break;
-                //вниз
-                case Direction.Down:
-                    if (_startPosY + step < _pictureHeight - planeHeight)
-                    {
-                        _startPosY += step;
-                    }
-                    break;
-            }
-        }
-
-        /// <summary>
-        /// Отрисовка автомобиля
-        /// </summary>
-        /// <param name="g"></param>
-        public override void DrawTransport(Graphics g)
-        {
-            Pen pen = new Pen(Color.Black);
-            Brush brMain = new SolidBrush(MainColor);
-
-            // Рисуем корпус
-
-            g.FillEllipse(brMain, _startPosX + 80, _startPosY + 50, 50, 20);
-            g.FillEllipse(brMain, _startPosX, _startPosY + 50, 40, 20);
-            g.FillRectangle(brMain, _startPosX + 10, _startPosY + 50, 100, 20);
-            g.FillEllipse(brMain, _startPosX + 10, _startPosY + 30, 20, 60);
-            g.FillEllipse(brMain, _startPosX + 70, _startPosY, 20, 120);
-        }
-    }
-}
