+++ WindowsFormsCars/WindowsFormsCars/WindowsFormsPlane.csproj
@@ -46,7 +46,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="APlane.cs" />
     <Compile Include="Direction.cs" />
     <Compile Include="FormPlane.cs">
       <SubType>Form</SubType>
@@ -54,10 +53,8 @@
     <Compile Include="FormPlane.Designer.cs">
       <DependentUpon>FormPlane.cs</DependentUpon>
     </Compile>
-    <Compile Include="ITransport.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Plane.cs" />
     <Compile Include="RadarPlane.cs" />
     <EmbeddedResource Include="FormPlane.resx">
       <DependentUpon>FormPlane.cs</DependentUpon>
